requirements.txt
mysql_connector_repackaged==0.3.1
openpyxl==3.0.10
pandas==1.4.3
pwinput==1.0.2
